agente_pandero/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-cloud-bigquery
google-generativeai
google-genai